Add rich text and attached image
Still need to review if image should be cached and investigate how to delete an image without deleting the whole product. Leaving the image empty in "edit" form doesn't seem to do it.
Can the server pre-populate a file upload field in a form? I don't remember it being able to do so, but it's been a while.
Next section: https://guides.rubyonrails.org/getting_started.html#internationalization-i18n